+++ versionsupport_1_13_R2/src/main/java/com/andrei1058/bedwars/support/version/v1_13_R2/v1_13_R2.java
@@ -36,7 +36,6 @@ import net.minecraft.server.v1_13_R2.*;
 import org.bukkit.Color;
 import org.bukkit.Effect;
 import org.bukkit.Location;
-import org.bukkit.Particle;
 import org.bukkit.block.BlockState;
 import org.bukkit.block.data.type.Bed;
 import org.bukkit.command.Command;
@@ -698,7 +697,7 @@ public class v1_13_R2 extends VersionSupport {
     }
 
     @Override
-    public List<org.bukkit.block.Block> calculateExplosionBlocks(IArena arena, Entity source, Location explosionLocation, int radius, boolean fire, BiFunction<Location, org.bukkit.block.Block, Boolean> callback) {
+    public List<org.bukkit.block.Block> calculateExplosionBlocks(IArena arena, Entity source, Location explosionLocation, float radius, boolean fire, BiFunction<Location, org.bukkit.block.Block, Boolean> callback) {
         HashSet<org.bukkit.block.Block> blocks = new HashSet<>();
         org.bukkit.World bukkitWorld = explosionLocation.getWorld();
