+++ versionsupport_1_16_R2/src/main/java/com/andrei1058/bedwars/support/version/v1_16_R2/v1_16_R2.java
@@ -36,7 +36,6 @@ import com.mojang.datafixers.util.Pair;
 import net.minecraft.server.v1_16_R2.*;
 import org.bukkit.Color;
 import org.bukkit.Location;
-import org.bukkit.Particle;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockState;
 import org.bukkit.block.data.type.Bed;
@@ -690,7 +689,7 @@ public class v1_16_R2 extends VersionSupport {
     }
 
     @Override
-    public List<Block> calculateExplosionBlocks(IArena arena, Entity source, Location explosionLocation, int radius, boolean fire, BiFunction<Location, Block, Boolean> callback) {
+    public List<Block> calculateExplosionBlocks(IArena arena, Entity source, Location explosionLocation, float radius, boolean fire, BiFunction<Location, Block, Boolean> callback) {
         HashSet<Block> blocks = new HashSet<>();
         org.bukkit.World bukkitWorld = explosionLocation.getWorld();
